weavy::jit::async — graduate the async suspend/resume substrate (phase 3)
The proven mechanism moves out of the throwaway spike into the weavy crate
proper. weavy stays runtime-free: the driver uses only core::future; tokio
is a DEV-dependency (test executor + timers) only.

- stencils/async_ops.rs: weavy_async_{push,await,add,mul,done}, built by
  weavy/build.rs into jit::async_stencils with GUARANTEED tail calls
  (become via RUSTC_BOOTSTRAP on the stencil rustc only), so suspension is
  sound for arbitrary-length chains (one driver-owned frame).
- jit::async: the substrate — Ctx layout, Op, AsyncProgram + compile()
  (bakes each await's resume offset + index into the immediate stream),
  and AsyncExec: a core::future::Future driver with the blocked-await
  re-entry guard, plus first-class debuggability (trace: the suspension
  timeline; suspension(): live parked state with operand-stack snapshot).
  Comprehensive module rustdoc = the design (guard-stencil analogy, the
  state machine, why copy-and-patch can suspend at all).
- v1 operand type is i64 (canonical lane); Value-generalization and an IR
  suspend node are the documented follow-ups. The suspend/resume PROTOCOL
  is the reusable part.
- Deleted weavy-async-spike and removed it from the workspace.

5 async tests over a real tokio runtime inside weavy's suite (80 tests
total, all green): single await; ready-never-suspends; two-awaits park in
program order; independent awaits resolve concurrently (one park); parked
state inspectable ([30] on the stack). Default weavy + downstream JIT
consumers (gingembre spike, vix-wire) build clean. Clippy clean.
