--- weavy-async-spike/Cargo.toml
@@ -1,21 +1,0 @@
-# SPIKE — throwaway. Proves the ONE thing that gates lowering vix's
-# demand-driven evaluation to weavy: can a copy-and-patch chain suspend at an
-# await and resume, driven by real Rust async? If yes, JIT'd vix nodes become
-# pollable futures that suspend on real network I/O (vox streams). Keeper = the
-# ANSWER; delete the crate once it graduates into weavy::jit::async.
-
-[package]
-name = "weavy-async-spike"
-version = "0.0.0"
-edition = "2024"
-publish = false
-description = "SPIKE: copy-and-patch suspend/resume over real tokio (weavy async crux)."
-
-[dependencies]
-weavy = { path = "../weavy", features = ["jit"] }
-
-[build-dependencies]
-copypatch = { path = "../copypatch", features = ["build"] }
-
-[dev-dependencies]
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync", "time"] }
