--- weavy-async-spike/build.rs
@@ -1,76 +1,0 @@
-//! Extract the async stencils' machine code + continuation relocations, exactly
-//! like the intop lane's build.rs. Emits `$OUT_DIR/async_stencils.rs`.
-
-use std::{env, fs, path::PathBuf};
-
-fn main() {
-    use std::fmt::Write as _;
-    let manifest = PathBuf::from(env::var("CARGO_MANIFEST_DIR").unwrap());
-    let out = PathBuf::from(env::var("OUT_DIR").unwrap());
-    let generated = out.join("async_stencils.rs");
-    let src = manifest.join("stencils/async_ops.rs");
-    println!("cargo:rerun-if-changed=build.rs");
-    println!("cargo:rerun-if-changed={}", src.display());
-
-    let target_os = env::var("CARGO_CFG_TARGET_OS").unwrap_or_default();
-    let target_arch = env::var("CARGO_CFG_TARGET_ARCH").unwrap_or_default();
-    let native = matches!(
-        (target_os.as_str(), target_arch.as_str()),
-        ("macos", "aarch64") | ("linux", "x86_64")
-    );
-    if !native {
-        let empty = ["PUSH", "AWAIT", "ADD", "MUL", "DONE"]
-            .iter()
-            .map(|n| format!("pub const {n}: &[u8] = &[];\npub const {n}_CONT: &[usize] = &[];\n"))
-            .collect::<String>();
-        fs::write(&generated, empty).expect("write empty stencils");
-        return;
-    }
-
-    let target = env::var("TARGET").expect("TARGET set by cargo");
-    let obj = out.join("async_stencils.o");
-    let rustc = env::var("RUSTC").unwrap_or_else(|_| "rustc".to_string());
-
-    // Guaranteed tail calls via `become` on stable (RUSTC_BOOTSTRAP=1) — the
-    // discipline that makes suspension sound (no live C-stack state across a
-    // continuation, so unwinding to suspend loses nothing).
-    unsafe { env::set_var("RUSTC_BOOTSTRAP", "1") };
-    let tailcall = copypatch::extract::compile_object(&rustc, &[], &src, &obj, &target, true);
-    if !tailcall {
-        assert!(
-            copypatch::extract::compile_object(&rustc, &[], &src, &obj, &target, false),
-            "rustc failed to compile async stencils"
-        );
-    }
-    println!(
-        "cargo:warning=async stencils compiled with {}",
-        if tailcall { "become" } else { "stable call" }
-    );
-
-    let bytes = fs::read(&obj).expect("read stencil object");
-    let symbols = [
-        "weavy_push",
-        "weavy_await",
-        "weavy_add",
-        "weavy_mul",
-        "weavy_done",
-    ];
-    let get = |sym: &str| copypatch::extract::extract_stencil(&bytes, &symbols, sym, "weavy_cont");
-
-    let mut s = String::new();
-    s.push_str("// @generated by build.rs from stencils/async_ops.rs.\n");
-    for (name, sym) in [
-        ("PUSH", "weavy_push"),
-        ("AWAIT", "weavy_await"),
-        ("ADD", "weavy_add"),
-        ("MUL", "weavy_mul"),
-        ("DONE", "weavy_done"),
-    ] {
-        let st = get(sym);
-        writeln!(s, "pub const {name}: &[u8] = &{:?};", st.bytes).unwrap();
-        if name != "DONE" {
-            writeln!(s, "pub const {name}_CONT: &[usize] = &{:?};", st.cont_relocs).unwrap();
-        }
-    }
-    fs::write(&generated, s).expect("write async stencils");
-}
