+++ weavy/Cargo.toml
@@ -23,5 +23,10 @@ copypatch = { workspace = true, features = ["build"], optional = true }
 default = []
 jit = ["dep:copypatch"]
 
+# The async lane's driver uses only `core::future` — no runtime dependency.
+# tokio is dev-only, to provide an executor + timers for the suspend tests.
+[dev-dependencies]
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "time"] }
+
 [lints]
 workspace = true
