+++ weavy/build.rs
@@ -14,15 +14,83 @@ fn main() {
     {
         let out = PathBuf::from(env::var("OUT_DIR").unwrap());
         let generated = out.join("weavy_stencils.rs");
+        let async_generated = out.join("weavy_async_stencils.rs");
         let target_os = env::var("CARGO_CFG_TARGET_OS").unwrap_or_default();
         let target_arch = env::var("CARGO_CFG_TARGET_ARCH").unwrap_or_default();
 
         if native_copy_patch_target(&target_os, &target_arch) {
             emit_native(&out, &generated);
+            emit_async_native(&out, &async_generated);
         } else {
             emit_empty(&generated);
+            emit_async_empty(&async_generated);
         }
     }
+}
+
+/// Async lane stencils: the suspend/resume substrate for the demand-driven
+/// evaluator. Compiled with GUARANTEED tail calls (`become`, via
+/// RUSTC_BOOTSTRAP on the stencil rustc only) so suspension is sound for
+/// arbitrary-length chains — the whole chain runs in one driver-owned frame.
+#[cfg(feature = "jit")]
+const ASYNC_SYMBOLS: &[&str] = &[
+    "weavy_async_push",
+    "weavy_async_await",
+    "weavy_async_add",
+    "weavy_async_mul",
+    "weavy_async_done",
+];
+
+#[cfg(feature = "jit")]
+fn emit_async_empty(generated: &Path) {
+    let mut s = String::new();
+    for n in ["PUSH", "AWAIT", "ADD", "MUL", "DONE"] {
+        s.push_str(&format!(
+            "pub const {n}: &[u8] = &[];\npub const {n}_CONT: &[usize] = &[];\n"
+        ));
+    }
+    fs::write(generated, s).unwrap();
+}
+
+#[cfg(feature = "jit")]
+fn emit_async_native(out: &Path, generated: &Path) {
+    println!("cargo:rerun-if-changed=stencils/async_ops.rs");
+
+    let target = env::var("TARGET").expect("TARGET set by cargo");
+    let obj = out.join("weavy_async_stencils.o");
+    let src = Path::new("stencils/async_ops.rs");
+    let rustc = env::var("RUSTC").unwrap_or_else(|_| "rustc".to_string());
+
+    // `become` needs `explicit_tail_calls`; RUSTC_BOOTSTRAP=1 unlocks it on the
+    // stable toolchain for THIS spawned rustc only (never downstream crates).
+    // At -O the bytes are identical to a plain call LLVM already TCOs; `become`
+    // is the correctness guard that suspension can't silently regress.
+    unsafe { env::set_var("RUSTC_BOOTSTRAP", "1") };
+    let tailcall = compile_object(&rustc, &[], src, &obj, &target, true);
+    if !tailcall {
+        assert!(
+            compile_object(&rustc, &[], src, &obj, &target, false),
+            "rustc failed to compile Weavy async stencils"
+        );
+    }
+
+    let bytes = fs::read(&obj).unwrap();
+    let mut s = String::new();
+    s.push_str("// Generated by build.rs from stencils/async_ops.rs (async suspend lane).\n");
+    for (name, sym) in [
+        ("PUSH", "weavy_async_push"),
+        ("AWAIT", "weavy_async_await"),
+        ("ADD", "weavy_async_add"),
+        ("MUL", "weavy_async_mul"),
+        ("DONE", "weavy_async_done"),
+    ] {
+        let st = extract_stencil(&bytes, ASYNC_SYMBOLS, sym, "weavy_cont");
+        emit(&mut s, name, &format!("`{sym}` async stencil."), &st);
+        if name != "DONE" {
+            emit_cont(&mut s, &format!("{name}_CONT"), name, &st);
+        }
+    }
+    fs::write(generated, s).unwrap();
 }
 
 #[cfg(feature = "jit")]
