+++ weavy/src/jit/mod.rs
@@ -76,10 +76,18 @@ pub mod stencils {
     include!(concat!(env!("OUT_DIR"), "/weavy_stencils.rs"));
 }
 
+/// Machine code + continuation relocations for the ASYNC (suspend/resume) lane,
+/// generated by `build.rs` from `stencils/async_ops.rs`. Empty off the native
+/// copy-and-patch targets.
+pub mod async_stencils {
+    include!(concat!(env!("OUT_DIR"), "/weavy_async_stencils.rs"));
+}
+
 // NB: module docs live INSIDE these files (`//!`); an outer `///` here would make rustdoc
 // resolve the merged doc's intra-doc links in THIS scope and break them.
 pub mod debug;
 pub mod dwarf;
+pub mod r#async;
 
 /// Whether this build can allocate and run native copy-and-patch code.
 pub const NATIVE_COPY_PATCH_AVAILABLE: bool = cfg!(any(
